src/api/database.py
import mysql.connector


class DatabaseManager:
    def __init__(self, host, port, user, password, database):
        self.connection = mysql.connector.connect(
            host=host,
            port=port,
            user=user,
            password=password,
            database=database
        )

        self.cursor = self.connection.cursor(dictionary=True)


    def execute_query(self, query, params=None, fetch_all=True):
        if params:
            self.cursor.execute(query, params)
        else:
            self.cursor.execute(query)
        
        return self.cursor.fetchall() if fetch_all else self.cursor.fetchone()


    def execute_and_commit(self, query, params=None):
        if params:
            self.cursor.execute(query, params)
        else:
            self.cursor.execute(query)

        self.connection.commit()
        